move QgsCurvePolygon tests from QgsGeometry
